+++ processor/src/main/kotlin/io/mcarle/konvert/processor/konvert/KonverterCodeGenerator.kt
@@ -65,7 +65,11 @@ object KonverterCodeGenerator {
                         codeBuilder.addFunction(
                             funBuilder = FunSpec.builder(konvertData.mapFunctionName)
                                 .addModifiers(KModifier.OVERRIDE)
-                                .returns(konvertData.targetTypeReference.toTypeName())
+                                // when mapping into an existing target instance, the function may return Unit
+                                .returns(
+                                    konvertData.mapKSFunctionDeclaration.returnType?.toTypeName()
+                                        ?: konvertData.targetTypeReference.toTypeName()
+                                )
                                 .addParameters(konvertData.mapKSFunctionDeclaration.parameters.map {
                                     val builder = ParameterSpec.builder(
                                         name = it.name!!.asString(),
@@ -133,6 +137,8 @@ object KonverterCodeGenerator {
                     target = konvertData.targetType,
                     paramName = konvertData.paramName,
                     targetClassImportName = targetClassImportName,
+                    targetParamName = konvertData.targetParamName,
+                    returnsTargetParam = konvertData.returnsTargetParameter,
                 ),
                 visibilityContext = MappingVisibilityContext.Declaration(konvertData.mapKSFunctionDeclaration),
                 additionalSourceParameters = konvertData.additionalParameters